Fold classic spec twins into zoneless specs (issue #2423 part c)
Merge the behavioral coverage from the four classic .component.spec.ts
twins into their .zoneless.spec.ts counterparts and delete the classic
files, leaving one spec per component (matching the already-migrated
find-view/browse-* pattern). Setup is normalized to configureZoneless();
test bodies are preserved verbatim (all already used provideZoneless() +
TestBed.tick()/settleResource(), the sanctioned zoneless helpers).

Components: right-panel, center-panel, image-viewer, label-view.
